add controls component, various fixes, refactor for component to function on its own
